+++ geopandas/tests/test_geodataframe.py
@@ -961,24 +961,24 @@ class TestDataFrame:
 
     def test_to_wkb(self):
         wkbs0 = [
-            (
+            (  # POINT (0 0)
                 b"\x01\x01\x00\x00\x00\x00\x00\x00\x00\x00\x00"
                 b"\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00"
-            ),  # POINT (0 0)
-            (
+            ),
+            (  # POINT (1 1)
                 b"\x01\x01\x00\x00\x00\x00\x00\x00\x00\x00"
                 b"\x00\xf0?\x00\x00\x00\x00\x00\x00\xf0?"
-            ),  # POINT (1 1)
+            ),
         ]
         wkbs1 = [
-            (
+            (  # POINT (2 2)
                 b"\x01\x01\x00\x00\x00\x00\x00\x00\x00\x00"
                 b"\x00\x00@\x00\x00\x00\x00\x00\x00\x00@"
-            ),  # POINT (2 2)
-            (
+            ),
+            (  # POINT (3 3)
                 b"\x01\x01\x00\x00\x00\x00\x00\x00\x00\x00"
                 b"\x00\x08@\x00\x00\x00\x00\x00\x00\x08@"
-            ),  # POINT (3 3)
+            ),
         ]
         gs0 = GeoSeries.from_wkb(wkbs0)
         gs1 = GeoSeries.from_wkb(wkbs1)
@@ -1527,11 +1527,11 @@ def test_set_geometry_supply_colname(nybb2, geo_col_name):
     res = nybb2.set_geometry("centroid")
     assert res.active_geometry_name == "centroid"
     assert geo_col_name in res.columns
-
-    res2 = nybb2.set_geometry("centroid", drop=True)
-    # drop=True should preserve column name centroid
-    assert res2.active_geometry_name == "centroid"
-    assert geo_col_name not in res2.columns
+    with pytest.warns(FutureWarning, match="The drop keyword argument is deprecated"):
+        res2 = nybb2.set_geometry("centroid", drop=True)
+    # drop=True should preserve previous geometry col name (keep old behaviour)
+    assert res2.active_geometry_name == geo_col_name
+    assert "centroid" not in res2.columns
 
 
 @pytest.mark.parametrize("geo_col_name", ["geometry", "polygons"])
@@ -1542,7 +1542,15 @@ def test_set_geometry_supply_arraylike(nybb2, geo_col_name):
     res = nybb2.set_geometry(centroids)
     assert res.active_geometry_name == geo_col_name
     # drop should do nothing if the column already exists
-    res2 = nybb2.set_geometry(centroids, drop=True)
+    match_str = (
+        "The drop keyword argument is deprecated and has no effect when an array-like"
+        " is passed"
+    )
+    with pytest.warns(
+        FutureWarning,
+        match=match_str,
+    ):
+        res2 = nybb2.set_geometry(centroids, drop=True)
     assert res2.active_geometry_name == geo_col_name
 
     centroids = centroids.rename("centroids")
@@ -1552,7 +1560,11 @@ def test_set_geometry_supply_arraylike(nybb2, geo_col_name):
     assert res3.active_geometry_name == "centroids"
     assert geo_col_name in res3.columns
 
-    # Drop should remove previous active geometry colname
-    res4 = nybb2.set_geometry(centroids, drop=True)
+    # Drop should not remove previous active geometry colname for arraylike inputs
+    with pytest.warns(
+        FutureWarning,
+        match=match_str,
+    ):
+        res4 = nybb2.set_geometry(centroids, drop=True)
     assert res4.active_geometry_name == "centroids"
-    assert geo_col_name not in res4.columns
+    assert geo_col_name in res4.columns